kioskhero_web/requirements.txt
pip==23.2.1
django==4.2.4
djangorestframework==3.14.0
markdown==3.4.4
django-filter==23.2
whitenoise==6.5
gunicorn==21.2.0
psycopg[binary]==3.1.10
sentry-sdk==1.29.2
